Elements

Missing docstring.

Missing docstring for FLOWPanel.AbstractElement. Check Documenter's build log for details.

Constant Element

Missing docstring.

Missing docstring for FLOWPanel.ConstantSource. Check Documenter's build log for details.

FLOWPanel.U_constant_source — Function

Computes the velocity induced by a panel of vertices nodes[:, panel] and constant strength source strength on the targets targets. It adds the velocity at the i-th target to out[i].

Implementation of equations in Hess & Smith (1967).

FLOWPanel.phi_constant_source — Function

Computes the potential induced by a panel of vertices nodes[:, panel] and constant strength source strength on the targets targets. It adds the potential at the i-th target to out[i].

Implementation of equations in Hess & Smith (1967), p. 53.

Doublet Element

Missing docstring.

Missing docstring for FLOWPanel.ConstantDoublet. Check Documenter's build log for details.

FLOWPanel.U_constant_doublet — Function

Computes the velocity induced by a panel of vertices nodes[:, panel] and constant strength doublet strength on the targets targets. It adds the velocity at the i-th target to out[i].

FLOWPanel.phi_constant_doublet — Function

Computes the potential induced by a panel of vertices nodes[:, panel] and constant strength doublet strength on the targets targets. It adds the potential at the i-th target to out[i].

Implementation of equations in Hess & Smith (1967), using ϕµ = −µ*(n⋅∇ϕσ/σ).

FLOWPanel.U_semiinfinite_doublet — Function

Computes the velocity induced by a semi-infinite doublet panel of strength strength on the targets targets. The semi-infinite panel starts at the trailing edge between nodes[:, TE[1]] and nodes[:, TE[2]] and extends infinitely in the direction [d1, d2, d3]. It adds the potential at the i-th target to out[i].

FLOWPanel.phi_semiinfinite_doublet — Function

Computes the potential induced by a semi-infinite doublet panel of strength strength on the targets targets. The semi-infinite panel starts at the trailing edge between nodes[:, TE[1]] and nodes[:, TE[2]] and extends infinitely in the direction [d1, d2, d3]. It adds the potential at the i-th target to out[i].

Implementation of equations in Moran, J. (1984), Appendix F, p. 445.

Computes the potential induced by a semi-infinite horseshoe of strength strength on the targets targets. The semi-infinite horseshoe comes from ∞ to pa and goes out to ∞ from pb, with a bound vortex going from pa to pb. The semi-infinite directions of pa and pb are da and db, respectively. It adds the velocity at the i-th target to out[i].

DEFINITIONS

pa = nodes[:, TE[1]]

pb = nodes[:, TE[2]]

da = [da1, da2, da3]

db = [db1, db2, db3]

Implementation of equations in Moran, J. (1984), Appendix F, p. 445, splitting the non-planar semi-infinite panel into two planar sections.

Vortex Element

Missing docstring.

Missing docstring for FLOWPanel.VortexRing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FLOPanel.U_boundvortex. Check Documenter's build log for details.

FLOWPanel.U_vortexring — Function

U_vortexring(nodes::Matrix, panel::Array{Int}, strength::Real, targets::Matrix, out::Matrix; dot_with=nothing, closed_ring=true, cutoff=1e-14, offset=1e-8)

Computes the velocity induced by a vortex ring panel of vertices nodes[:, panel] and vortex strength strength on the targets targets. It adds the velocity at the i-th target to out[i].

FLOWPanel.U_semiinfinite_vortex — Function

Computes the velocity induced by a semi-infinite vortex starting at point [p1, p2, p3] going out in the unitary direction [d1, d2, d3] with vortex strength strength on the targets targets. It adds the velocity at the i-th target to out[i].

A negative strength makes it such that the vortex comes from infinite and ends at the point.

FLOWPanel.U_semiinfinite_horseshoe — Function

Computes the velocity induced by a semi-infinite horseshoe of strength strength on the targets targets. The semi-infinite horseshoe comes from ∞ to nodes[:, TE[1]] and goes out to ∞ from nodes[:, TE[2]], with a bound vortex going from nodes[:, TE[1]] to nodes[:, TE[2]]. The direction is of the semi-infinite sections is given by [d1, d2, d3]. It adds the velocity at the i-th target to out[i].

Computes the velocity induced by a semi-infinite horseshoe of strength strength on the targets targets. The semi-infinite horseshoe comes from ∞ to pa and goes out to ∞ from pb, with a bound vortex going from pa to pb. The semi-infinite directions of pa and pb are da and db, respectively. It adds the velocity at the i-th target to out[i].

DEFINITIONS

pa = nodes[:, TE[1]]

pb = nodes[:, TE[2]]

da = [da1, da2, da3]

db = [db1, db2, db3]

Missing docstring.

Missing docstring for FLOWPanel.U_semiinfinite_horseshoe. Check Documenter's build log for details.